chore: split tooltip style file
