+++ src/renderer/src/components/ui/tooltip/styles.ts
@@ -1,0 +1,9 @@
+export const tooltipStyle = {
+  content: [
+    "relative z-[101] border border-accent/10 bg-white px-2 py-1 text-foreground dark:bg-neutral-950",
+    "data-[state=closed]:animate-out data-[state=closed]:fade-out-0",
+    "rounded-lg text-sm",
+    "max-w-[75ch] select-text",
+    "drop-shadow data-[side=top]:shadow-tooltip-bottom data-[side=bottom]:shadow-tooltip-top dark:border-border",
+  ],
+}
